Limit legend texts width, and wrap them.
Note that this only affects legend texts managed by MapFish print, and
not legend icons labels, that are included inside the images returned
after the WMS GetLegendGraphic query.

Note also that these legend icons labels are always appended to the icon,
since mapfishapp pass the "LEGEND_OPTIONS=forceLables:on" GeoServer vendor
parameter. There is no mean to know exactly the resulting size of the
generated image, and no GeoServer parameter to restrict the label size,
therefore the two MapFish parameters that control the legend icons size
are let undefined (iconMaxWidth and iconMaxHeight).

This means that, whereas the legend texts managed by MapFish are wrapped
when they hit the legend max width, the image containing the icons and
the icons labels might overflow to the right of the page, and some part of
the label would be hidden, in case these labels are too long.