CAS - datadir
These modifications allows the retrieval of variables at runtime instead
of at compilation time, so that the end user are able to modify the
values into their properties file in the datadir instead of having to
recompile the whole solution.

Note: This PR shall be addressed at the same time as the same one on CAS
on the georchestra/georchestra repository.